Dark Mode Implementation & Nav Fixes
Works under a theme, to change the dark theme colours you need to change the p-colours in css.. it was made in a way where you just change one colour instead of every.

Work in progress with symbol changing when moon is clicked (moon turns in to sun, sun turns into moon).

The pages that are currently avaliable