Logout function + Basket uncommented
Done:
Uncommented basket as it now works functionally since login/logout has been added.
With logout, it appears on the navbar instead of login if a user has logged in. In addition, the welcome page will display their name in the 'Welcome to 13 Keys' -> this will also happen if the code sees that the user is logged in.
It appeared that a recent commit might of taken out the glowing on the navbar headers to display what page the user is currently on so added them back.